feat - added exception for existing tenant name
Added a TenantNameUnavailable exception that is triggered when
attempting to create a new account, or change the existing tenant name,
with a name that is already in use by another account.

Achieved this by setting a unique constraint on the tenant_name in the
database. When an attempt to use an existing value, SQL will throw a
data validation error, which is caught and a TenantNameUnavailable
exception is thrown to the caller.