fix - resolved tests not passing on dev machines
The issue was that the `spring.profiles.active` profile in the
`application.properties` file was set to use the production profile,
which expects a SQL database to be running on the local machine.

I have set the `spring.profile.active` value back to `test`, which uses
an in-memory H2 database. This is a short-term fix, and a new task has
been opened to come up with a more resilient method of switching
profiles depending on whether a developer is attempting to run tests
versus building a production-ready version of the app. See user story
CP-93 for more information.